extract mempool logic
